+++ src/fsck.fat.h
@@ -38,6 +38,7 @@
 #include <linux/msdos_fs.h>
 
 #include <stddef.h>
+#include <stdint.h>
 #include <endian.h>
 
 #define VFAT_LN_ATTR (ATTR_RO | ATTR_HIDDEN | ATTR_SYS | ATTR_VOLUME)
@@ -149,8 +150,8 @@ typedef struct _dos_file {
 } DOS_FILE;
 
 typedef struct {
-    unsigned long value;
-    unsigned long reserved;
+    uint32_t value;
+    uint32_t reserved;
 } FAT_ENTRY;
 
 typedef struct {
@@ -159,12 +160,12 @@ typedef struct {
     unsigned int fat_size;	/* unit is bytes */
     unsigned int fat_bits;	/* size of a FAT entry */
     unsigned int eff_fat_bits;	/* # of used bits in a FAT entry */
-    unsigned long root_cluster;	/* 0 for old-style root dir */
+    uint32_t root_cluster;	/* 0 for old-style root dir */
     loff_t root_start;
     unsigned int root_entries;
     loff_t data_start;
     unsigned int cluster_size;
-    unsigned long clusters;
+    uint32_t clusters;
     loff_t fsinfo_start;	/* 0 if not present */
     long free_clusters;
     loff_t backupboot_start;	/* 0 if not present */
@@ -180,7 +181,7 @@ extern void *mem_queue;
 
 /* value to use as end-of-file marker */
 #define FAT_EOF(fs)	((atari_format ? 0xfff : 0xff8) | FAT_EXTD(fs))
-#define FAT_IS_EOF(fs,v) ((unsigned long)(v) >= (0xff8|FAT_EXTD(fs)))
+#define FAT_IS_EOF(fs,v) ((uint32_t)(v) >= (0xff8|FAT_EXTD(fs)))
 /* value to mark bad clusters */
 #define FAT_BAD(fs)	(0xff7 | FAT_EXTD(fs))
 /* range of values used for bad clusters */
